+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `market_closed_borrowing_fee_base_factor`
   - `market_closed_borrowing_fee_above_optimal_usage_factor`
 - programs(store): Introduced flag `enable_market_closed_params` to control activation of closed-market configs.
+- programs(store): Added a new `MARKET_CONFIG_KEEPER` role with permissions to update market configs.
 - sdk(sdk): Added `try_deserialize_zero_copy_from_base64_with_options` utility function.
 - sdk(js): Added decode methods with `no_discriminator` option for JsMarket and JsPosition.
 - sdk(js): Added `add_create_orders_builder` function.